feat(mail): Use mail endpoint on message.fullstack.cash
